Big commit, but permit deletion now possible, as well as working with documents
Permits can be successfully deleted, with a pop-up that confirms the action before it takes place. Documents can be attached and saved. They can be removed as well. However, when a document exists and a new one is attached, a small error occurs. Yet, the document still attaches. Currently a WIP